Complete rework using Python Lex-Yacc (PLY) library
The entire structure of the project was redone. Instead of using a handmade lexer and parser, PLY provides a way to quickly setup new lexing tokens and parsing rules. Although challenging to setup at first, now that the framework is established, features can be implemented rather quickly. Additionally, the workshop.json file provides a way to type-check in the future. Variables and vectors can now be defined/created easily.