Containment operators, refactored all tokens, while loop, attributes
As a result of whitespace continuing to present an issue, I refactored the way tokens are parsed into a much more functional way. Each token is now matched directly as a string rather than extracting it from whitespace, which fixed issues with the `not in` and `in` operators. While loop functionality is extremely basic: it will repeat a block of code by looping if the condition is true. Attributes are ways to access properties more quickly (e.g. vector.x -> x component of(vector)).