Improve functions, attributes. Add methods and string support.
Functions now (kind of) resolve parameters. Methods are simply a combination of attribute and function call (so far only `array.push(x)` is added). Raw strings are declared with quotes while strings that have parameters are written as such: `{0}: {1}` and followed by a block or arg_list of the parameters.