Fix certain tokens in parsing, strings in arrays. Add disabled rules
Disabled rules can now be created by prepending rules with 'disabled'. Strings are not allowed in Overwatch arrays; however, to make better use of strings in arrays to utilize in for loops, they are kept internally. A number of attributes have been added such as eyepos, and many fixes have been made to token parsing.